Re-Implement old concept of AutoCombineRewardsThresholdTime; which appeared to be a failed attempt at delaying the AutoCombineDust() to occur after X minutes. That old algorithm was commented out and replaced with a 5 second wait. This concept also was flawed, in that it was going to be doing a sleep within the ProcessNewBlock() code. The first attempt was abandoned because it likely was blocking ProcessNewBlock for 15 minutes at a time, and no way to not block, if using the feature, for under a minute. The workaround (5 seconds) still wasn't desirable, as it would still lock up ProcessNewBlock for the 5 second block.
This new method changes that design from ThresholdTime, to Block
Frequency, and defaults to 15 blocks. Time can be adjusted per
implementation, to get a desired minute time based on the block
frequency of the coin parameters. If AutoCombineDust is enabled,
it will only check and attempt to combine dust if the block height
is a multiple of the Block Frequency. e.g. if set to 10, then every
10th block it will be executed. 100; then every 100th block.

Additionally, logic was added to provide a "one time" sweep; if
you don't want a regularly scheduled sweep but you want to do a
sweep at that given moment.  Specifying a block frequency of
zero will run AutoCombineDust() on the next block received, and
then turn it off.  It will be saved in the database as a one time
sweep; so it will then run every time you start up the wallet.
If you do not wish it to run on startup, you can turn off
autocombinerewards after the sweep was performed.

This can now be tailored by the user, based on their desired dust
cleanup threshold, and their expectation of frequency that they will
need to clean.