[PreRelease] Merge of v1.2 working branch to mainline (#40)
* Updated dnsseed nodes to be the new official nodes

Updated dnsseed nodes to be the new official nodes along with adding dns lookup nodes to faciliate with future changes without having to change the code.

* Fix PIV and zPIV references (#10)

* More zPiv and Piv eradication (#12)

* Fix PIV and zPIV references

* Fix more PIV and zPIV references

* Fix more PIV and zPIV references

* AutoCombineRewards fixes and Improvements (#16)

* Re-Implement old concept of AutoCombineRewardsThresholdTime; which
appeared to be a failed attempt at delaying the AutoCombineDust()
to occur after X minutes. That old algorithm was commented out and
replaced with a 5 second wait. This concept also was flawed, in
that it was going to be doing a sleep within the ProcessNewBlock()
code. The first attempt was abandoned because it likely was blocking
ProcessNewBlock for 15 minutes at a time, and no way to not block,
if using the feature, for under a minute. The workaround (5 seconds)
still wasn't desirable, as it would still lock up ProcessNewBlock
for the 5 second block.

This new method changes that design from ThresholdTime, to Block
Frequency, and defaults to 15 blocks. Time can be adjusted per
implementation, to get a desired minute time based on the block
frequency of the coin parameters. If AutoCombineDust is enabled,
it will only check and attempt to combine dust if the block height
is a multiple of the Block Frequency. e.g. if set to 10, then every
10th block it will be executed. 100; then every 100th block.

Additionally, logic was added to provide a "one time" sweep; if
you don't want a regularly scheduled sweep but you want to do a
sweep at that given moment.  Specifying a block frequency of
zero will run AutoCombineDust() on the next block received, and
then turn it off.  It will be saved in the database as a one time
sweep; so it will then run every time you start up the wallet.
If you do not wish it to run on startup, you can turn off
autocombinerewards after the sweep was performed.

This can now be tailored by the user, based on their desired dust
cleanup threshold, and their expectation of frequency that they will
need to clean.

* Merge in latest from Pivx plus fixes for threshold boundaries

* Update rpcserver.cpp

* Update based on PIVX feedback

* Fix -disablewallet crashes (#18)

* Update main.cpp

Add pwalletmain check and obfuScationPool message removal to cover when wallet is disabled (PIVX #817)

* Update net.cpp

Update to remove call to ThreadStakeMinter() when -disableWallet is in effect (PIVX PR #817)

* Update obfuscation.cpp

Return from UnlockCoins if there is no pwalletMain, to prevent crash with -disableWallet

* Layout and Logo Updates (#23)

* Logo and Color Changes

* Fixing QSlider coloring

* Bruno updates

* Add explorer address to seeders

* Autocombine to max possible & increase max possible (#26)

* Cleanup debug msg

* Update seednodes

* Remove more btc references from messages

* Update rpcwallet.cpp

* listunspent doc correction and typos

* Update rpcrawtransaction.cpp

* Fix Block Reward Calculations (#31)

When creating a new block, the blockreward being used was the block reward at the current chain height, not the block reward for the height that the chain would be once this new block is added. This inadvertently minted one extra block at the previous halving level.

This was not caught in consensus, because it also was calculating the block reward off the previous block, not the current block that it validates. So two wrongs didn't make it right, but made it functional.

Lastly, because the payee checks during consensus only check to make sure the block creator isn't trying to short change the masternodes, that code didn't care that the block creator was giving twice as much to the masternode then it should have... on those crossover blocks.

This corrects both sides, having the block creation look ahead in the rewards table for the reward relevant to the block being created, not the previous block created. Likewise the consensus looks up the current height of the block being validated, rather than the previous height.

This addresses PR #30

* More BlockReward fix (#32)

* Fix FillBlockPayee lookup of BlockValue and MasternodePayment

* Get ExpectedMint for current block

* Get Reward from the right height

* Ignore overmints from the past

* Disablewallet (#33)

* Update main.cpp

Add pwalletmain check and obfuScationPool message removal to cover when wallet is disabled (PIVX #817)

* Update net.cpp

Update to remove call to ThreadStakeMinter() when -disableWallet is in effect (PIVX PR #817)

* Update obfuscation.cpp

Return from UnlockCoins if there is no pwalletMain, to prevent crash with -disableWallet

* Masternode Drift to Percent (#34)

* Block reward (#35)

* Fix FillBlockPayee lookup of BlockValue and MasternodePayment

* Get ExpectedMint for current block

* Masternode Drift to Percent

* Get Reward from the right height

* Ignore overmints from the past

* typo

* Several Fixes 

Fix block value on reward switch blocks
- Fix Block Payee to reflect the proper Block Value at reward switch blocks.
- Fix ExpectedMint to get the expected mint for the current block, not the previous block.
- Get the block reward from the correct height
- Add code to ignore the overmint on the first halving, so we don't have any problems loading the block history.

Convert Masternode Drift from a masternode count, to a percentage of the locked coins.

Fix the initial Proof of Stake difficulty calculation.

Fix testnet generation

* Fix Backwards Compatability of old Autocombine Rewards (#37)

* Do not autocombine 2 UTXOs or less

As the autocombine will result in no less than 2 UTXOs (The combined UTXO and the change UTXO), we don't want to combine if we only find 2 or less UTXOs.

* Update Spork Keys, Disable Zerocoin (#38)

* Update removing Qt files

* Delete Gamblecoin.cflags

* Delete Gamblecoin.config

* Delete Gamblecoin.creator

* Delete Gamblecoin.includes

* Delete Gamblecoin.cxxflags

* Delete Gamblecoin.files

* Update Version

* staking and mining code rework and cleanup

* Finishing touches on miner changes

* Also fix the block height for budget payments

* Fix Typos

* More seednode tweaks